Refactor 'excludeJobIds' Serialization for Spark Job Compatibility (#155)
This commit changes the serialization format of the 'excludeJobIds' field in the session document. Previously serialized as an array (e.g., ["id1", "id2"]), this field is now formatted as a comma-separated string ("id1,id2"). This adjustment is made to ensure compatibility with plugins that currently do not support array parsing. This simplification aids in seamless integration and easier data handling in downstream processes.

Testing done:
* modified unit test
Signed-off-by: Kaituo Li <kaituo@amazon.com>