Can integration tests against the docker cluster
The Python script for integration tests was updated to run queries against the docker cluster.
The required indices are created as part of the script. The queries for the Python script were
likely out of date. These have been updated when the fix for the query was obvious.

There are still 6 tests that fail.

Signed-off-by: Norman Jordan <norman.jordan@improving.com>